feat: Nearby page (기능 수정) , Detailpage (시간 텍스트, 모달창 밑줄 수정)
